requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.7.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.1.1
gunicorn
pytz
cryptography
flask_cors
MySQLdb
pillow
razorpay
reportlab